requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
comtypes==1.1.11
DateTime==4.5
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gTTS==2.2.4
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
playsound==1.3.0
protobuf==4.21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycairo==1.21.0
pygame==2.1.2
PyGObject==3.42.2
pyparsing==3.0.9
python-decouple==3.6
pyttsx3==2.90
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SpeechRecognition==3.8.1
uritemplate==4.1.1
urllib3==1.26.10
zope.interface==5.4.0
